update spec & arc docs (executor) (#33)
* upd insert executor

* upd insert executor

* upd insert executor

* upd select executor

+++ internal/planner/create_table.go
@@ -101,7 +101,7 @@ func getUkParams(ukDefs []*definition.ConstraintUniqueKeyDef, colIndexMap map[st
 		uniqueKeyParams = append(uniqueKeyParams, &executor.IndexParam{
 			Name:         ukDef.KeyName,
 			ColName:      ukDef.Column.ColName,
-			SecondaryKey: uint(idx),
+			SecondaryKey: uint16(idx),
 		})
 	}
 	return uniqueKeyParams, nil
